Adds changes required to convert AST into 3AC
This CR adds all changes that we needed to convert the AST we
have so far to 3AC. It includes some temporary work arounds, such as,
using a Vec<AstNode> instead of a proper AstNode variant to collect all
the AST nodes inside of the function given in the test cases in stage 4.
The current method being used to walk the AST is recursive - which is
less than ideal and not useable in real world scenarios. I should look
into whether it is possible to implement it iteratively.