Refactors symbol_table module
This CR refactors the symbol table module to represent the global
symbol table as an enum instead of a struct. The two symbol table
enum variants allow us to model the behavior of our symbol table
correctly as our symbol table should be defunct upon the creation
of tiny code.