Fixes logic bug in tiny code generation
This CR fixes a logic bug in tiny code generation which was
leading us to insert incorrect register values for 3AC temporaries
in the register maps maintained by the tiny module, for Math binary
expressions - add, sub, mul, div. Instead of caching the tiny register
(in the register map, keyed against 3AC temporaries) where we were
actually storing the result of the binary expression, we were creating
and caching a new register.