Overhauls symbol table.
The CR contains a number of changes in the symbol table module. A
very few of those changes have to do with some cosmetic stuff around
"safe unwrapping", which I wanted to get rid of after getting some code
feedback. Rest of the changes were made to allow the addition of
function names as symbols.

I did not try to create a `Symbol` enum representing function and
data symbols. Instead I changed the existing Symbol type to the new
DataSymbol type and created a new FunctionSymbol type. Not sure if
that is the best way to model data but we'll see.

I feel like I have left the codebase in a state of higher entropy.
There's a few other changes that probably need to happen along with
the introduction of function names as symbols. I'm most annoyed by how
I have an Identifier type that seems to have no relation whatsoever with
my symbol table. There should be a better way to structure identifiers
such that they borrow from the symbol table and have a back reference
to the corresponding entry in the symbol table.

Overall this change supports the bigger task of adding support for
function calls.