Adds the plumbing required to support an AST node for functions
This commit mainly modifies 3 modules, all under symbol_table. First, it
modifies the symbol module to add a new FunctionDataSymbol type. This
new type will be used to represent data symbols - locals and parameters,
in functions. The type still needs to be merged (made a variant of) the
DataSymbol type because that is what all other statements - assignment,
addition etc. use.

Second, the commit modifies the scope module. It modifies the structure of
Scope::FunctionScope and adds methods to add/lookup a function data symbol to
a scope in the existing scope tree. The commit also improves the error
handling in this module and adds unit tests to the module.

Third, the commit also makes some small modifications to the root
symbol_table module, mainly adding a method to add a function data
symbol to the symbol table. Had to create some TODOs regarding testing
that will need to be handled later.

I attempted to modify the grammar/parser code so that I could at least
get a functional AST node for functions generated, that could then be
used for code gen. However, I figured out that the FunctionDataSymbol
type will need to be merged with the DataSymbol type rather late.
Solcing that issue is going to be a pre-requisite to generate an AST
node for functions. The parsing code for parameter list for functions
also needs to be modified so that it returns a Vec<NumTypes>
(essentially the arguments part of a function's signature).

Most of the code gen is commented out and the code compiles. The grammar
file currently contains a lot of dummy code, however.