Adds more changes in preparation for enabling Function AST nodes
This CR adds more changes that are required to enable the creation of
function AST nodes. The key realization that I had in this CR was that
the stack frame slot information associated to parameters and locals in
functions needs to live with the symbols and not as part of the function
scope.

As mentioned in the previous CR, I also "merged" the data symbols (now
non func scoped symbols) and func data symbols (now func scoped
symbols), because either of these symbols can occur in the different
expressions supported in Micro. Both of these type of symbols are now
variants of an umbrella symbol enum called `Symbol`. The `Identifier`
node in the ast_node module should now consist of a `Symbol` instead of
a `NonFunctionScopedSymbol` (previously called `DataSymbol`).

I also fixed a bug in the symbol lookup by name method, which wasn't
looking for symbols recursively up the scope hierarchy. In order to fix
this bug I had to change the structure of the `ScopeTree` and the
different `Scope` variants themselves. Each `Scope` variant now has a
shared pointer to its parent (which is stored in a vector of shared
pointers in `ScopeTree`).