Adds 3AC code gen for functions and function calls
This CR adds the capability to generate 3AC code for functions and
function calls. There seems to be some edge cases around functions
returning VOID which will need to be handled since there are multiple
cases in 3AC codegen where we unwrap results of expressions, which up
till now were always guaranteed to be `Some`.

There's also a decent amount of unit testing TODOs that have accumulated
that I plan to complete in the next CR or two.