Adds tiny code generation support for functions
Changes in this CR get the tiny code generation for functions to work.
In order to get the tiny code gen to work I had to poke just a little
bit at the innards of -

1. `ast_node::Stmt::Return` - changing the `Return` statement to
contain an `Assignment` instead of just an expression.

2. `microc.lalrpop` - accompanying the change to `Return` statement to
treat it as an `Assignment` rather than just an expression, I
modified the syntax-directed translation of functions to always create a
parameter (read a function scope symbol), which the `Return` statement
now assigns to.

3. `symbol::data::FunctionScopedSymbolType` - changing the `Parameter`
variant to contain a count of the total number of parameters in the
function, which is used during tiny code gen to address function
parameters as positive offsets to the frame pointer. The reason I did
not have to do something similar for local variables was because
currently function parameters follow the "Pascal" order on the stack
(for no good reason) and I can possibly change this in the future and
reduce some complexity.

4. `three_addr_code_ir::three_address_code` - modifying the 3AC visitor
logic for `Return` statement based on above changes and removing
"now" unused variants - `StoreFuncReturnValI` and `StoreFuncReturnValF`
of the `ThreeAddressCode` enum.

The rest of the changes were made to the `tiny` module itself.