Refactors `BinaryExprOperand` type into respective int and float types
This CR removes the existing `BinaryExprOperand` type and replaces it
with 2 new types - `BinaryExprOperandI` and `BinaryExprOperandF`. The
change corrects a few different issues in the 3AC code module.
1. Different 3AC operations  that are int and float specific now contain
   binary expression operands of the correct types, i.e., the lhs and
   rhs of a 3AC code such as `AddI` will be `BinaryExprOperandI`.
2. The `result` of a `CodeObject`, which used to be a
   `BinaryExprOperand` is now a new type `LValue` - an enum with
   variants `LValueI` and `LValueF`. The `result` of a `CodeObject` cannot
   be an RValue. By using the existing `BinaryExprOperand` type we were not
   enforcing this rule.

The introduction of int and float specific binary expression operands
snowballed into a pretty big refactor. However, it improved the
correctness of the constructs we use and will simplify a few things when
we perform liveness analysis.