Adds some testing and cfg printing improvements
This CR adds a unit test to test conversion of a bb function to a
liveness decorated control flow graph. I'm not too happy with how the
strategy to calculate liveness is so strongly coupled with the control
flow grpah type. Maybe there is a better way to decouple the 2.

The CR also improves the printing of the CFG by aligning the GEN, KILL,
IN and OUT sets for all 3AC nodes.