Fixes the `ThreeAddressCode` data model
This CR fixes the `ThreeAddressCode` data model by way of fixing the
data type of the field used by a number of its variants. Before this CR
a number of 3AC codes were using the `BinaryExprOperand` type (which has
been renamed to `RValue` in this CR). `BinaryExprOperand` could have
been an lvalue or an rvalue, however no operand of any of the 3ACs,
except for STOREI and STOREF can actually be an rvalue.

This clean up will make the register allocation logic simpler as we can
now simply perform the register allocation logic/operations on lvalues -
which can only be identifiers or temporaries, which makes sense.

I had to comment of tiny code generation for now as there were numerous
new errors introduced because of the refactor introduced in this CR. I
will likely need to do a fairly big refactor in the tiny code generation
logic as post register allocation it should become much more straight
forward than it is today.