Adds changes for generating register allocated 3AC
This CR adds the code for generating register allocated 3AC using the
`RegisterFile` construct that we defined in the previous CR. In addition
to the changes for generating the register allocated 3AC, in this CR I
had to address some missed gaps in the `RegisterFile` type, the
major one being handling spilling of temporary variables. The way I'm
handling spills of temporaries is by creating associated
`FunctionScopedSymbols` for them when we do need to spill them. Handling
spilling of temporaries in this way should simplify the story of
generating tiny code for spills/stores of temporaries to stack memory.

Some of the methods in the methods in the `RegisterFile` type need to be
improved as they are reaching into the guts of the type where we can
introduce a layer of abstraction using additional methods. For instance,
in the `ensure_register` method I am directly accessing the
`tempi_to_func_scoped_symbol_map` field as a hashmap, thus making it
hard for my future self to refactor things if required.

I also still need to add unit tests for the `RegisterFile` type.