Gets tiny code gen with register allocation partially working
This CR implements tiny code generation for register allocated 3AC. Tiny
code generation now works partially. There are a few caveats that will
need to be addressed before it can be considered as fully working.

1. Function calls do not work -
I am not currently pushing all the registers to the stack before
function calls. The code to spill all the dirty registers carrying
global vars is also untested.

2. Even just a single function does not quite work yet -
At the beginning of each function a given amount of stack space needs to
be reserved. The amount of space is specified as part of the `link`
command. Currently I'm storing the amount of stack space that may be
used during register allocation (in the `stack_top_idx` field of the
`RegisterFile` type) but I'm not passing it through to the function that
generates tiny code from register allocated 3AC.

Other than the above 2 functional gaps, I seriously need to add unit
tests -

1. `register_alloc/mod.rs`
This should test the end to end conversion of a list of liveness
decorated 3AC into register allocated 3AC.

2. `register_alloc/types.rs`
This should test the functionality of the different primitive structs in
the module, but mostly `RegisterFile`.

Code is in a bit of a chaotic static. A little clean up will be helpful
:).
.