Gets register allocation working 95 percent
With this CR register allocation works pretty much completely except for
programs which contains function calls. The reason register allocation
is not working in such programs is because I am not storing the register
containing the "return" variable (which is passed in as the 1st
parameter to a function implicitly) back to the stack, before returning
from a function.

The underlying reason for this behavior is that the
"return" variable is currently not being marked as a LIVE variable out
of the return statement when I'm initializing GEN and KILL sets before
liveness analysis.

I'll need to figure out how/where to plumb the right information such
that the return value is treated as LIVE in basic blocks containing
return statements in a function.

Overall, good push this weekend to get stuff almost fully working. In
hindsight most of the trouble this weekend happened because I did not
believe that BOTTOM-UP register allocation only works at the basic block
level. But I learnt a thing or 2 :).

Unit tests coverage is still not great, even though I did add a few
tests this weekend. Maybe I'll work on them after fixing the function
call issue.